Fix path bug
