+++ src/lib/mgylabs/utils/config.py
@@ -4,6 +4,8 @@ import os
 import sys
 import threading
 
+from .version import VERSION
+
 config_sema = threading.BoundedSemaphore()
 
 LOCALAPPDATA = os.getenv("LOCALAPPDATA")
@@ -18,9 +20,9 @@ if is_development_mode():
     MGCERT_PATH = "../data/mgcert.json"
     USER_DATA_PATH = "../data"
 else:
-    CONFIG_PATH = f"{LOCALAPPDATA}\\Mulgyeol\\Mulgyeol MK Bot\\data\\config.json"
-    MGCERT_PATH = f"{LOCALAPPDATA}\\Mulgyeol\\Mulgyeol MK Bot\\data\\mgcert.json"
-    USER_DATA_PATH = f"{LOCALAPPDATA}\\Mulgyeol\\Mulgyeol MK Bot\\data"
+    CONFIG_PATH = f"{LOCALAPPDATA}\\Mulgyeol\\{VERSION.product_name}\\data\\config.json"
+    MGCERT_PATH = f"{LOCALAPPDATA}\\Mulgyeol\\{VERSION.product_name}\\data\\mgcert.json"
+    USER_DATA_PATH = f"{LOCALAPPDATA}\\Mulgyeol\\{VERSION.product_name}\\data"
 
 
 class SettingItem:
@@ -121,40 +123,6 @@ class Settings:
             self.save()
 
 
-class Version:
-    def __init__(self, version_str, commit):
-        version_str = version_str.split("-")
-        self.commit = commit
-        self.base_version = version_str[0]
-        self.tuple_version = tuple(self.base_version.split("."))
-        self.canary = (
-            True
-            if (
-                len(version_str) > 1
-                and version_str[1] == "beta"
-                and self.commit != None
-            )
-            else False
-        )
-        self.tag = f"v{self.base_version}"
-
-    def is_release_build(self):
-        return (not is_development_mode()) and (self.commit is not None)
-
-    def is_canary(self) -> bool:
-        return self.canary
-
-    def __str__(self) -> str:
-        if is_development_mode():
-            return "Dev"
-        if self.commit == None:
-            return f"{self.base_version} Test Mode"
-        if self.is_canary():
-            return f"{self.base_version} Canary"
-        else:
-            return f"{self.base_version} Stable"
-
-
 def invoke():
     if os.path.isfile(CONFIG_PATH):
         with open(CONFIG_PATH, "rt", encoding="utf-8") as f:
@@ -187,19 +155,9 @@ def add_data(key, value):
         json.dump(TOKEN, f, indent=4, ensure_ascii=False)
 
 
-def get_mkbot_version():
-    try:
-        with open("../info/version.json", "rt", encoding="utf-8") as f:
-            d = json.load(f)
-        return Version(d["version"], d["commit"])
-    except Exception:
-        return None
-
-
 def get_discriminator(key):
     return hashlib.sha1(key.encode()).hexdigest()
 
 
 CONFIG = Settings(invoke())
-VERSION = get_mkbot_version()
 DISCRIMINATOR = get_discriminator(CONFIG.discordToken)
